submissions/requirements.txt
torch
transformers>=4.41.0
pillow
gradio
pydantic==2.10.6
sentencepiece